🧹 remove standard lib dependency from core layer ✨ default layout

+++ src/cumpk80.cpp
@@ -1,5 +1,3 @@
-#include <cstdint>
-
 #include "core/umpk80.hpp"
 #include "core/disassembler.hpp"
 
@@ -14,9 +12,9 @@ extern "C" {
     } UMPK80_Instruction_t;
 
     typedef struct UMPK80_I8080DisassembleResult {
-        uint16_t address;
-        uint8_t bytesCount;
-        uint8_t bytes[3];
+        u16 address;
+        u8 bytesCount;
+        u8 bytes[3];
         const UMPK80_Instruction_t* instruction;
         bool eof;
     } UMPK80_I8080DisassembleResult_t;
@@ -43,48 +41,48 @@ extern "C" {
     UMPK80_t UMPK80_Create();
     void     UMPK80_Free(UMPK80_t umpk);
 
-    void    UMPK80_PortIOSetInput(UMPK80_t umpk, uint8_t data);
-    uint8_t UMPK80_PortIOGetInput(UMPK80_t umpk);
+    void    UMPK80_PortIOSetInput(UMPK80_t umpk, u8 data);
+    u8 UMPK80_PortIOGetInput(UMPK80_t umpk);
 
-    uint8_t UMPK80_PortIOGetOutput(UMPK80_t umpk);
+    u8 UMPK80_PortIOGetOutput(UMPK80_t umpk);
 
     void    UMPK80_Tick(UMPK80_t umpk);
     void    UMPK80_Stop(UMPK80_t umpk);
     void    UMPK80_Restart(UMPK80_t umpk);
 
-    void    UMPK80_KeyboardPressButton(UMPK80_t umpk, uint8_t key);
-    void    UMPK80_KeyboardReleaseButton(UMPK80_t umpk, uint8_t key);
+    void    UMPK80_KeyboardPressButton(UMPK80_t umpk, u8 key);
+    void    UMPK80_KeyboardReleaseButton(UMPK80_t umpk, u8 key);
 
-    uint8_t UMPK80_DisplayGetDigit(UMPK80_t umpk, int digit);
-    void    UMPK80_LoadOS(UMPK80_t umpk, const uint8_t* os);
+    u8 UMPK80_DisplayGetDigit(UMPK80_t umpk, int digit);
+    void    UMPK80_LoadOS(UMPK80_t umpk, const u8* os);
 
-    void    UMPK80_LoadProgram(UMPK80_t umpk, const uint8_t* program, uint16_t programSize, uint16_t dstAddress);
+    void    UMPK80_LoadProgram(UMPK80_t umpk, const u8* program, u16 programSize, u16 dstAddress);
 
-    uint16_t UMPK80_CpuProgramCounter(UMPK80_t umpk);
-    void     UMPK80_CpuSetProgramCounter(UMPK80_t umpk, uint16_t value);
-    uint16_t UMPK80_CpuStackPointer(UMPK80_t umpk);
-    uint8_t  UMPK80_CpuGetRegister(UMPK80_t umpk, UMPK80_Register reg);
-    void     UMPK80_CpuSetRegister(UMPK80_t umpk, UMPK80_Register reg, uint8_t data);
-    void     UMPK80_CpuJump(UMPK80_t umpk, uint16_t adr);
-    void     UMPK80_CpuCall(UMPK80_t umpk, uint16_t adr);
+    u16 UMPK80_CpuProgramCounter(UMPK80_t umpk);
+    void     UMPK80_CpuSetProgramCounter(UMPK80_t umpk, u16 value);
+    u16 UMPK80_CpuStackPointer(UMPK80_t umpk);
+    u8  UMPK80_CpuGetRegister(UMPK80_t umpk, UMPK80_Register reg);
+    void     UMPK80_CpuSetRegister(UMPK80_t umpk, UMPK80_Register reg, u8 data);
+    void     UMPK80_CpuJump(UMPK80_t umpk, u16 adr);
+    void     UMPK80_CpuCall(UMPK80_t umpk, u16 adr);
 
-    uint8_t UMPK80_GetRegister(UMPK80_t umpk, UMPK80_Register reg);
-    void    UMPK80_SetRegister(UMPK80_t umpk, UMPK80_Register reg, uint8_t value);
+    u8 UMPK80_GetRegister(UMPK80_t umpk, UMPK80_Register reg);
+    void    UMPK80_SetRegister(UMPK80_t umpk, UMPK80_Register reg, u8 value);
 
-    uint16_t UMPK80_GetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair);
-    void     UMPK80_SetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair, uint16_t value);
+    u16 UMPK80_GetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair);
+    void     UMPK80_SetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair, u16 value);
 
-    uint8_t UMPK80_MemoryRead(UMPK80_t umpk, uint16_t adr);
-    void    UMPK80_MemoryWrite(UMPK80_t umpk, uint16_t adr, uint8_t data);
+    u8 UMPK80_MemoryRead(UMPK80_t umpk, u16 adr);
+    void    UMPK80_MemoryWrite(UMPK80_t umpk, u16 adr, u8 data);
 
-    const UMPK80_Instruction_t* UMPK80_GetInstruction(uint8_t code);
+    const UMPK80_Instruction_t* UMPK80_GetInstruction(u8 code);
 
-    UMPK80_I8080Disassembler_t UMPK80_CreateI8080Disassembler(const uint8_t* memory, size_t size);
+    UMPK80_I8080Disassembler_t UMPK80_CreateI8080Disassembler(const u8* memory, u64 size);
     void                       UMPK80_FreeI8080Disassembler(UMPK80_I8080Disassembler_t disasm);
 
     UMPK80_I8080DisassembleResult_t UMPK80_I8080DisassemblerDisassemble(UMPK80_I8080Disassembler_t disasm);
     void UMPK80_I8080DisassemblerReset(UMPK80_I8080Disassembler_t disasm);
-    uint16_t UMPK80_I8080DisassemblerPG(UMPK80_I8080Disassembler_t disasm);
+    u16 UMPK80_I8080DisassemblerPG(UMPK80_I8080Disassembler_t disasm);
 }
 
 
@@ -98,15 +96,15 @@ void UMPK80_Free(UMPK80_t umpk) {
 
 static Umpk80* inst(UMPK80_t umpk) { return (Umpk80*)umpk; }
 
-void UMPK80_PortIOSetInput(UMPK80_t umpk, uint8_t data) {
+void UMPK80_PortIOSetInput(UMPK80_t umpk, u8 data) {
     inst(umpk)->port5InSet(data); 
 }
 
-uint8_t UMPK80_PortIOGetInput(UMPK80_t umpk, uint8_t data) {
+u8 UMPK80_PortIOGetInput(UMPK80_t umpk, u8 data) {
     return inst(umpk)->port5InGet(); 
 }
 
-uint8_t UMPK80_PortIOGetOutput(UMPK80_t umpk) {
+u8 UMPK80_PortIOGetOutput(UMPK80_t umpk) {
     return inst(umpk)->port5OutGet(); 
 }
 
@@ -122,35 +120,35 @@ void UMPK80_Restart(UMPK80_t umpk) {
     inst(umpk)->restart();
 }
 
-void UMPK80_KeyboardPressButton(UMPK80_t umpk, uint8_t key) {
+void UMPK80_KeyboardPressButton(UMPK80_t umpk, u8 key) {
     inst(umpk)->pressKey((KeyboardKey)key);
 }
 
-void UMPK80_KeyboardReleaseButton(UMPK80_t umpk, uint8_t key) {
+void UMPK80_KeyboardReleaseButton(UMPK80_t umpk, u8 key) {
     inst(umpk)->releaseKey((KeyboardKey)key);
 }
 
-uint8_t UMPK80_DisplayGetDigit(UMPK80_t umpk, int digit) {
+u8 UMPK80_DisplayGetDigit(UMPK80_t umpk, int digit) {
     return inst(umpk)->getDisplayDigit(digit);
 }
 
-void UMPK80_LoadOS(UMPK80_t umpk, const uint8_t* os) {
+void UMPK80_LoadOS(UMPK80_t umpk, const u8* os) {
     inst(umpk)->loadOS(os);
 }
 
-uint16_t UMPK80_CpuProgramCounter(UMPK80_t umpk) {
+u16 UMPK80_CpuProgramCounter(UMPK80_t umpk) {
     return inst(umpk)->getCpu().getProgramCounter();
 }
 
-void UMPK80_CpuSetProgramCounter(UMPK80_t umpk, uint16_t value) {
+void UMPK80_CpuSetProgramCounter(UMPK80_t umpk, u16 value) {
     inst(umpk)->getCpu().setProgramCounter(value);    
 }
 
-uint16_t UMPK80_CpuStackPointer(UMPK80_t umpk) {
+u16 UMPK80_CpuStackPointer(UMPK80_t umpk) {
     return inst(umpk)->getCpu().getStackPointer();
 }
 
-uint8_t  UMPK80_CpuGetRegister(UMPK80_t umpk, UMPK80_Register reg) {
+u8  UMPK80_CpuGetRegister(UMPK80_t umpk, UMPK80_Register reg) {
     auto& cpu = inst(umpk)->getCpu();
     
     switch (reg) {
@@ -170,14 +168,14 @@ uint8_t  UMPK80_CpuGetRegister(UMPK80_t umpk, UMPK80_Register reg) {
     }
 }
 
-void UMPK80_CpuSetRegister(UMPK80_t umpk, UMPK80_Register reg, uint8_t data) {
+void UMPK80_CpuSetRegister(UMPK80_t umpk, UMPK80_Register reg, u8 data) {
     auto& cpu = inst(umpk)->getCpu();
 
     switch (reg) {
         case PC_LOW:  cpu.setProgramCounter((cpu.getProgramCounter() & 0xFF00) | data);
-        case PC_HIGH: cpu.setProgramCounter((cpu.getProgramCounter() & 0x00FF) | (((uint16_t)data) << 8));
+        case PC_HIGH: cpu.setProgramCounter((cpu.getProgramCounter() & 0x00FF) | (((u16)data) << 8));
         case SP_LOW:  cpu.setStackPointer((cpu.getStackPointer() & 0xFF00) | data);
-        case SP_HIGH: cpu.setStackPointer((cpu.getStackPointer() & 0x00FF) | (((uint16_t)data) << 8));
+        case SP_HIGH: cpu.setStackPointer((cpu.getStackPointer() & 0x00FF) | (((u16)data) << 8));
         case L:       cpu.setRegister(Cpu::Register::L, data);
         case H:       cpu.setRegister(Cpu::Register::H, data);
         case E:       cpu.setRegister(Cpu::Register::E, data);
@@ -190,48 +188,48 @@ void UMPK80_CpuSetRegister(UMPK80_t umpk, UMPK80_Register reg, uint8_t data) {
     }
 }
 
-void UMPK80_CpuJump(UMPK80_t umpk, uint16_t adr) {
+void UMPK80_CpuJump(UMPK80_t umpk, u16 adr) {
     inst(umpk)->getCpu().forceJump(adr);
 }
 
-void UMPK80_CpuCall(UMPK80_t umpk, uint16_t adr) {
+void UMPK80_CpuCall(UMPK80_t umpk, u16 adr) {
     inst(umpk)->getCpu().forceCall(adr);
 }
 
-uint8_t UMPK80_GetRegister(UMPK80_t umpk, UMPK80_Register reg) {
+u8 UMPK80_GetRegister(UMPK80_t umpk, UMPK80_Register reg) {
     return inst(umpk)->getRegister((Umpk80::Register)reg);
 }
 
-void UMPK80_SetRegister(UMPK80_t umpk, UMPK80_Register reg, uint8_t value) {
+void UMPK80_SetRegister(UMPK80_t umpk, UMPK80_Register reg, u8 value) {
     inst(umpk)->setRegister((Umpk80::Register)reg, value);
 }
 
-uint16_t UMPK80_GetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair) {
+u16 UMPK80_GetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair) {
     return inst(umpk)->getRegisterPair((Umpk80::RegisterPair)regPair);
 }
 
-void UMPK80_SetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair, uint16_t value) {
+void UMPK80_SetRegisterPair(UMPK80_t umpk, UMPK80_RegisterPair regPair, u16 value) {
     inst(umpk)->setRegisterPair((Umpk80::RegisterPair)regPair, value);
 }
 
-uint8_t UMPK80_MemoryRead(UMPK80_t umpk, uint16_t adr) {
+u8 UMPK80_MemoryRead(UMPK80_t umpk, u16 adr) {
     return inst(umpk)->getBus().memoryRead(adr);
 }
 
-void UMPK80_MemoryWrite(UMPK80_t umpk, uint16_t adr, uint8_t data) {
+void UMPK80_MemoryWrite(UMPK80_t umpk, u16 adr, u8 data) {
     inst(umpk)->getBus().memoryWrite(adr, data);
 }
 
-void UMPK80_LoadProgram(UMPK80_t umpk, const uint8_t* program, uint16_t programSize, uint16_t dstAddress) {
+void UMPK80_LoadProgram(UMPK80_t umpk, const u8* program, u16 programSize, u16 dstAddress) {
     inst(umpk)->getBus().loadRam(program, programSize, dstAddress);
 }
 
-const UMPK80_Instruction_t* UMPK80_GetInstruction(uint8_t code) {
+const UMPK80_Instruction_t* UMPK80_GetInstruction(u8 code) {
     return (UMPK80_Instruction_t*) &Disassembler::getInstruction(code);
 }
 
-UMPK80_I8080Disassembler_t UMPK80_CreateI8080Disassembler(const uint8_t *memory,
-                                                          size_t size) {
+UMPK80_I8080Disassembler_t UMPK80_CreateI8080Disassembler(const u8 *memory,
+                                                          u64 size) {
     return (UMPK80_I8080Disassembler_t) new Disassembler(memory, size);
 }
 
@@ -252,6 +250,6 @@ void UMPK80_I8080DisassemblerReset(UMPK80_I8080Disassembler_t disasm) {
     ((Disassembler*)disasm)->reset();
 }
 
-uint16_t UMPK80_I8080DisassemblerPG(UMPK80_I8080Disassembler_t disasm) {
+u16 UMPK80_I8080DisassemblerPG(UMPK80_I8080Disassembler_t disasm) {
     ((Disassembler*)disasm)->getPgCounter();
 }
